+++ test/Test/DecisionSpec.hs
@@ -1,22 +1,27 @@
-module Test.DecisionSpec where
+module Test.DecisionSpec
+       ( decisionSemigroupAssoc
+       , decisionMonoidMempty
 
-import Hedgehog (MonadGen, forAll, property, (===))
-import Test.Tasty (TestTree)
-import Test.Tasty.Hedgehog (testProperty)
+         -- * Gen
+       , genDecision
+       ) where
+
+import Hedgehog (MonadGen, Property, forAll, property, (===))
 
 import Summoner.Decision (Decision)
 
 import qualified Hedgehog.Gen as Gen
 
-test_decisionSemigroupAssoc :: [TestTree]
-test_decisionSemigroupAssoc = pure $ testProperty "Decision Semigroup:Assoc" $ property $ do
+
+decisionSemigroupAssoc :: Property
+decisionSemigroupAssoc = property $ do
     x <- forAll genDecision
     y <- forAll genDecision
     z <- forAll genDecision
     (x <> y) <> z === x <> (y <> z)
 
-test_decisionMonoidMempty :: [TestTree]
-test_decisionMonoidMempty = pure $ testProperty "Decision Monoid:Mempty" $ property $ do
+decisionMonoidMempty :: Property
+decisionMonoidMempty = property $ do
     x <- forAll genDecision
     x <> mempty === x
     mempty <> x === x