+++ test/golden/fullProject/src/Prelude.hs
@@ -1,0 +1,7 @@
+-- | Uses [relude](https://hackage.haskell.org/package/relude) as default Prelude.
+
+module Prelude
+       ( module Relude
+       ) where
+
+import Relude